Move navigation types file
